Package static data and type definitions into separate submodules.
