+++ smwe_rom/src/graphics/gfx_file/mod.rs
@@ -1,16 +1,19 @@
+mod data;
+
 use std::{
     convert::TryInto,
     fmt,
     fmt::{Display, Formatter},
 };
 
+pub(crate) use data::GFX_FILES_META;
 use nom::{bytes::complete::take, combinator::map_parser, multi::many1, IResult};
 
 use crate::{
     compression::lc_lz2,
     error::{DecompressionError, GfxFileParseError, RomError},
     graphics::color::{Abgr1555, Rgba32},
-    snes_utils::{addr::AddrSnes, rom::Rom, rom_slice::SnesSlice},
+    snes_utils::rom::Rom,
 };
 
 #[derive(Copy, Clone, Debug)]
@@ -153,58 +156,3 @@ impl GfxFile {
 // -------------------------------------------------------------------------------------------------
 
 pub const N_PIXELS_IN_TILE: usize = 8 * 8;
-#[rustfmt::skip]
-pub(crate) static GFX_FILES_META: [(TileFormat, SnesSlice); 0x34] = [
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08D9F9), 2104)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08E231), 2698)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08ECBB), 2199)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08F552), 2603)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08FF7D), 2534)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x098963), 2569)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09936C), 2468)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x099D10), 2375)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09A657), 2378)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09AFA1), 2676)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09BA15), 2439)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09C39C), 2503)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09CD63), 2159)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09D5D2), 2041)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09DDCB), 2330)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09E6E5), 2105)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09EF1E), 2193)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09F7AF), 2062)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x09FFBD), 2387)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0A8910), 2616)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0A9348), 1952)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0A9AE8), 2188)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AA374), 1600)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AA9B4), 2297)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AB2AD), 2359)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0ABBE4), 1948)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AC380), 2278)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0ACC66), 2072)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AD47E), 2058)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0ADC88), 2551)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AE67F), 1988)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AEE43), 2142)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AF6A1), 2244)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0AFF65), 2408)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0B88CD), 2301)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0B91CA), 2331)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0B9AE5), 2256)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BA3B5), 2668)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BAE21), 2339)),
-    (TileFormat::Tile4bpp, SnesSlice::new(AddrSnes(0x0BB744), 2344)),
-    (TileFormat::Tile2bpp, SnesSlice::new(AddrSnes(0x0BC06C), 1591)),
-    (TileFormat::Tile2bpp, SnesSlice::new(AddrSnes(0x0BC6A3), 1240)),
-    (TileFormat::Tile2bpp, SnesSlice::new(AddrSnes(0x0BCB7B), 1397)),
-    (TileFormat::Tile2bpp, SnesSlice::new(AddrSnes(0x0BD0F0), 1737)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BD7B9), 2125)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BE006), 2352)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BE936), 2127)),
-    (TileFormat::Tile2bpp, SnesSlice::new(AddrSnes(0x0BF185), 566)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BF3BB), 1093)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x0BF800), 1293)),
-    (TileFormat::Tile4bpp, SnesSlice::new(AddrSnes(0x088000), 16320)),
-    (TileFormat::Tile3bpp, SnesSlice::new(AddrSnes(0x08BFC0), 6713)),
-];
